Moved ConvertParseModelToUmlModel() method to model class UmlWorkspace
Update issue 12
Created model package and moved class UmlWorkspace in there.  As well as refactoring to give it the method ConvertParseModelToUmlModel() which should always have been in there anyway.

Note: before we start mapping to wx shapes, it seems we have the 
 * 'ParseModel' the parse model (our original model created by the command line tool pynsource)
 * 'UmlModel' the UmlWorkspace  model (more advanced model, subsumes the old model)
the latter contains lots of juicy info e.g. 
 * dict of classname => shape entries  # self.classnametoshape = {}
 * list of (classname, parentclassname) tuples  # self.associations_generalisation = []
 * list of (rhs, lhs) tuples  # self.associations_composition = []
 * UmlGraph()   # self.graph
and the graph is full of GraphNode objects, and associations, which are ultimately what we layout.

- arguably the dict of classname => shape entries is a rendering related mapping which shouldn't be here?

- I think we need the association mappings because the graph object we use only has simple associations not the level of detail we need.  Perhaps re-examing this and rely only on a      beefed up graph class instead - in order to simplify things?