Need to rework sibling colouring algorithm - hexmvc1.pyns is not colouring nicely and it should.
Update issue 16
Need to rework sibling colouring algorithm - hexmvc1.pyns is not colouring nicely and it should.
Want siblings of different parents to be a different colour.
Want all root nodes to be the same color, unless they have siblings, in which case the root should match the kid sibling colour.