Delete codesee-arch-diagram.yml
Hi Andy,

Josh here, from CodeSee. Sorry you ran into trouble with our GitHub Action. 

In order to uninstall our action, you need to remove the workflow file -- which I'm doing in this PR. Hope that helps!

In the meantime, I'm looking into why our action is hanging on this repo. Thank you so much for letting us know, I'm excited to get to the bottom of it!

Sorry again for the inconvenience,
Josh